新增成绩查询、搜索提示、更新成绩功能
BUG:搜索提示前端页面组件显示不出

+++ shop-service/shop-service-order/src/main/java/com/hand/demo/order/infra/feign/client/ProductFeignClient.java
@@ -15,5 +15,5 @@ import org.springframework.web.bind.annotation.RequestMethod;
 public interface ProductFeignClient {
 
     @RequestMapping(value = "/product/{id}", method = RequestMethod.GET)
-    Product findById(@PathVariable("id") Integer id);
+    Product findById(@PathVariable("id") Long id);
 }
